Move all labels to a more centralised location: src/apps/mydb/elements/labels/
The only file breaking this is the src/components/UserLabels.js.
It is used in src/components/navigation/UserAuth.js so I considered it more consistent to leave it where it is.
