+++ app/packs/src/apps/mydb/elements/details/reactions/ReactionDetails.js
@@ -6,8 +6,8 @@ import {
 } from 'react-bootstrap';
 import SvgFileZoomPan from 'react-svg-file-zoom-pan-latest';
 import { findIndex } from 'lodash';
-import ElementCollectionLabels from 'src/apps/mydb/elements/list/ElementCollectionLabels';
-import ElementAnalysesLabels from 'src/apps/mydb/elements/list/ElementAnalysesLabels';
+import ElementCollectionLabels from 'src/apps/mydb/elements/labels/ElementCollectionLabels';
+import ElementAnalysesLabels from 'src/apps/mydb/elements/labels/ElementAnalysesLabels';
 import ElementActions from 'src/stores/alt/actions/ElementActions';
 import DetailActions from 'src/stores/alt/actions/DetailActions';
 import LoadingActions from 'src/stores/alt/actions/LoadingActions';
