+++ app/packs/src/apps/mydb/elements/labels/ComputedPropLabel.js
