+++ app/packs/src/apps/mydb/elements/list/ElementsTableSampleEntries.js
@@ -2,14 +2,14 @@ import React, { Component } from 'react';
 import { Table, Button, Tooltip, OverlayTrigger, Label } from 'react-bootstrap';
 import PropTypes from 'prop-types';
 import ElementCheckbox from 'src/apps/mydb/elements/list/ElementCheckbox';
-import ElementCollectionLabels from 'src/apps/mydb/elements/list/ElementCollectionLabels';
-import ElementAnalysesLabels from 'src/apps/mydb/elements/list/ElementAnalysesLabels';
-import ElementReactionLabels from 'src/apps/mydb/elements/list/ElementReactionLabels';
-import ElementWellplateLabels from 'src/apps/mydb/elements/list/ElementWellplateLabels';
-import GenericElementLabels from 'src/components/generic/GenericElementLabels';
+import ElementCollectionLabels from 'src/apps/mydb/elements/labels/ElementCollectionLabels';
+import ElementAnalysesLabels from 'src/apps/mydb/elements/labels/ElementAnalysesLabels';
+import ElementReactionLabels from 'src/apps/mydb/elements/labels/ElementReactionLabels';
+import ElementWellplateLabels from 'src/apps/mydb/elements/labels/ElementWellplateLabels';
+import GenericElementLabels from 'src/apps/mydb/elements/labels/GenericElementLabels';
 import PubchemLabels from 'src/components/pubchem/PubchemLabels';
-import ChemrepoLabels from 'src/apps/mydb/elements/list/ChemrepoLabels';
-import ComputedPropLabel from 'src/components/computedProps/ComputedPropLabel';
+import ChemrepoLabels from 'src/apps/mydb/elements/labels/ChemrepoLabels';
+import ComputedPropLabel from 'src/apps/mydb/elements/labels/ComputedPropLabel';
 import ArrayUtils from 'src/utilities/ArrayUtils';
 import ElementContainer from 'src/apps/mydb/elements/list/ElementContainer';
